Move media functions into util
