+++ api/models/Media.js
@@ -1,17 +1,7 @@
 /* eslint-disable camelcase */
 import GetImageSize from '../services/GetImageSize'
 import request from 'request'
-import { merge } from 'lodash'
-import { pick } from 'lodash/fp'
-
-export const VALID_MEDIA_TYPES = [
-  'file',
-  'image',
-  'gdoc',
-  'video'
-]
-
-export const DEFAULT_MEDIA_TYPE = 'file'
+import { createAndAddSize } from './media/util'
 
 module.exports = bookshelf.Model.extend({
   tableName: 'media',
@@ -135,28 +125,3 @@ module.exports = bookshelf.Model.extend({
     return Promise.resolve()
   }
 })
-
-const createAndAddSize = function (attrs) {
-  const url = attrs.type === 'image'
-    ? attrs.url
-    : attrs.type === 'video'
-      ? attrs.thumbnail_url
-      : null
-
-  if (url) {
-    return GetImageSize(url).then(dimensions =>
-      Media.create(merge({}, attrs, pick(['width', 'height'], dimensions))))
-  }
-
-  return Media.create(attrs)
-}
-
-export function getMediaTypeFromMimetype (mimetype) {
-  let baseMimetype = mimetype && mimetype.split('/')[0]
-
-  // NOTE: This doesn't account for the currently unsupportd
-  // legacy special cases of 'gdoc' or 'video'
-  return VALID_MEDIA_TYPES.includes(baseMimetype)
-    ? baseMimetype
-    : DEFAULT_MEDIA_TYPE
-}
